Company is able to update users, user roles, and user account statuses

+++ app/Http/Controllers/Company/UsersController.php
@@ -30,10 +30,7 @@ class UsersController extends Controller
      */
     public function index()
     {
-        $roles = Role::all();
-        $permissions = Permission::all();
-
-        $user_account_statuses = UserAccountStatus::all();
+        $this->authorize('view', User::class);
 
         $users = User::with([
             'userAccountStatus',
@@ -42,10 +39,7 @@ class UsersController extends Controller
         ])->paginate(20);
 
         return view('dashboard/company/users/index', compact([
-            'users', 
-            'user_account_statuses',
-            'roles',
-            'permissions'
+            'users',
         ]));
     }
 
@@ -57,27 +51,46 @@ class UsersController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('update', User::class);
+
         $validatedData = $request->validate([
             'first_name' => 'required|string|max:255',
             'last_name' => 'required|string|max:255',
             'phone_number' => 'required|max:255',
             'email' => 'required|string|email|max:255|unique:users',
-            'user_account_status_id' => 'nullable|integer|exists:user_account_statuses,id',
             'password' => 'required|string|min:8|confirmed',
         ]);
 
-        User::create([
+        $user = User::create([
             'first_name' => $validatedData['first_name'],
             'last_name' => $validatedData['last_name'],
             'phone_number' => $validatedData['phone_number'],
             'email' => $validatedData['email'],
             'password' => Hash::make($validatedData['password']),
-            'user_account_status_id' => $validatedData['user_account_status_id'],
-            'status_by' => 1, // Auth::user()->id,
-            'status_date' => now(),
         ]);
 
-        return back();
+        return redirect()->route('company.users.edit', ['user' => $user])->with('success', 'User Added Successfully');
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  \App\Models\User  $user
+     * @return \Illuminate\Http\Response
+     */
+    public function edit(User $user)
+    {
+        $this->authorize('update', User::class);
+
+        $user = $user->load(['userAccountStatus', 'roles']);
+        $roles = Role::all();
+        $user_account_statuses = UserAccountStatus::all();
+        
+        return view('dashboard/company/users/edit', compact(
+            'user',
+            'roles',
+            'user_account_statuses'
+        ));
     }
 
     /**
@@ -89,12 +102,13 @@ class UsersController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        $this->authorize('update', User::class);
+
         $validatedData = $request->validate([
             'first_name' => 'required|string|max:255',
             'last_name' => 'required|string|max:255',
             'phone_number' => 'required|max:255',
             'email' => 'required|string|email|max:255|unique:users,email,'. $user->id,
-            'user_account_status_id' => 'nullable|integer|exists:user_account_statuses,id',
         ]);
 
         $user->update([
@@ -102,10 +116,9 @@ class UsersController extends Controller
             'last_name' => $validatedData['last_name'],
             'phone_number' => $validatedData['phone_number'],
             'email' => $validatedData['email'],
-            'user_account_status_id' => $validatedData['user_account_status_id'],
         ]);
 
-        return back();
+        return back()->with('success', 'User Details Updated Successfully');
     }
 
     /**
@@ -116,7 +129,10 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete', User::class);
+
         $user->delete();
-        return back();
+
+        return back()->with('success', 'User Deleted Successfully');
     }
 }
